requirements.txt
click
cryptodome
